+++ app/src/main/java/org/sopt/dosopttemplate/home/HomeFragment.kt
@@ -1,10 +1,23 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.home
 
-import androidx.lifecycle.ViewModel
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import org.sopt.dosopttemplate.FriendAdapter
 import org.sopt.dosopttemplate.Model.Friend
+import org.sopt.dosopttemplate.R
+import org.sopt.dosopttemplate.databinding.FragmentHomeBinding
 
-class HomeViewModel : ViewModel() {
-    val mockFriendList = listOf<Friend>(
+class HomeFragment : Fragment() {
+    private var _binding: FragmentHomeBinding? = null
+    private val binding: FragmentHomeBinding
+        get() = requireNotNull(_binding) { "FragmentHome 바인딩 객체가 생성되지 않았습니다" }
+
+    // 임시 데이터
+    private val mockFriendList = listOf<Friend>(
         Friend(
             profileImage = R.drawable.iv_profile,
             name = "전성기 시절 파트장",
@@ -51,4 +64,27 @@ class HomeViewModel : ViewModel() {
             aboutMe = "표정 풀자",
         ),
     )
+
+    private val viewModel by viewModels<HomeViewModel>()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val friendAdapter = FriendAdapter(requireContext())
+        binding.rvFriends.adapter = friendAdapter
+        friendAdapter.setFriendList(viewModel.mockFriendList)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
 }
