+++ app/src/main/java/org/sopt/dosopttemplate/home/HomeViewModel.kt
@@ -1,21 +1,11 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.home
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
+import androidx.lifecycle.ViewModel
 import org.sopt.dosopttemplate.Model.Friend
-import org.sopt.dosopttemplate.databinding.FragmentHomeBinding
+import org.sopt.dosopttemplate.R
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding: FragmentHomeBinding
-        get() = requireNotNull(_binding) { "FragmentHome 바인딩 객체가 생성되지 않았습니다" }
-
-    // 임시 데이터
-    private val mockFriendList = listOf<Friend>(
+class HomeViewModel : ViewModel() {
+    val mockFriendList = listOf<Friend>(
         Friend(
             profileImage = R.drawable.iv_profile,
             name = "전성기 시절 파트장",
@@ -62,27 +52,4 @@ class HomeFragment : Fragment() {
             aboutMe = "표정 풀자",
         ),
     )
-
-    private val viewModel by viewModels<HomeViewModel>()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        val friendAdapter = FriendAdapter(requireContext())
-        binding.rvFriends.adapter = friendAdapter
-        friendAdapter.setFriendList(viewModel.mockFriendList)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
